WIP for BltCommandTester interactive tester and testWizard test.
I've tried to get this to work and not sure how to proceed. I've added all of the code from the Symfony CommandTester class and modified the parts as noted to try and use BLT. In order to do some of this, I've needed to pass in a configuration object.

Errors aren't really produced, but going to the sandbox directory that is listed in the PHPUnit output and looking at the `project.yml` file, it doesn't look like the command is run.

I'm not sure how to step through any of this to see what is actually happening in order to debug it.